add a Coordinator interface to allow mocking
